refactor: Make transcript depend on enrollment
