--- backend/src/main/java/com/tkpm/sms/domain/repository/TranscriptRepository.java
@@ -1,12 +1,0 @@
-package com.tkpm.sms.domain.repository;
-
-import com.tkpm.sms.application.dto.response.enrollment.AcademicTranscriptDto;
-import com.tkpm.sms.domain.model.Enrollment;
-import com.tkpm.sms.domain.model.Student;
-
-import java.util.List;
-
-public interface TranscriptRepository {
-
-    AcademicTranscriptDto getAcademicTranscript(Student student, List<Enrollment> enrollments);
-}
